==> scripts/05_viz_thermal_pref.r <==
library(tidyverse)


# read in and go from here
source("scripts/load_data_settings.r")


# filter to species we have
coefout <- filter(coefout, gen_spp %in% subsite_substrate_long$gen_spp)
View(coefout)
# plot data ---------------------------------------------------------------


color2 <- "#782391"

coefout %>% 
  filter(!is.na(gen_spp)) |>
  mutate(gen_spp = forcats::fct_reorder(gen_spp, BO21_tempmax_bdmin_mean) ) %>%
  
  ggplot(aes(x=gen_spp)) +
  
  
  geom_point(aes(y=BO21_tempmax_bdmean_q5),  color = color2, alpha=.5, size=2)+
  geom_point(aes(y=BO21_tempmax_bdmean_q95), color = color2, alpha=.5, size=2)+
  geom_point(aes(y=BO21_tempmax_bdmean_mean), color = color2, alpha=1, size=3)+
  geom_segment(aes(xend=gen_spp,
                   y=BO21_tempmax_bdmean_q5,
                   yend=BO21_tempmax_bdmean_q95), color = color2, alpha=.5)+
#  annotate(geom="text",
#           x=3, y=30, 
#           hjust=-0.1, vjust=0.2,
#           label = 
#             "\n Max temp at mean depth",
#           color = color2,
#           fontface="bold",
#           size=5) +
  
  labs(x=NULL, y= expression(paste("Water Temperature in ", "\u00b0C"))) +
  theme_classic() +
  theme(#plot.margin = margin(l=25,b=5,unit="pt"),
    axis.text.x = element_text(angle = -90, hjust = 0))+
  geom_hline(yintercept=c(15.3, 17.5), linetype='dashed', color=c('turquoise', 'green')) + #Pre 2004 95% quantile for 15m and 25m
  geom_hline(yintercept=c(17.9, 19.6), color=c('turquoise', 'green')) + #Post 2004 95% quantile for 15m and 25m
  theme(axis.text.x = element_text(face = "italic")) +
  theme(axis.text=element_text(size=18),
        axis.title=element_text(size=18))

ggsave("figures/thermal_preference_ranges.jpg")
